+++ e2e/src/app.e2e-spec.ts
@@ -7,8 +7,8 @@ describe('workspace-project App', () => {
     page = new AppPage();
   });
 
-  it('should display welcome message', () => {
+  it('should display "Buttons" in appbar', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('Welcome to pip-webui2-buttons-example!');
+    expect(page.getParagraphText()).toEqual('Buttons');
   });
 });