+++ projects/pip-webui2-buttons/src/lib/fab-speed-dial/fab-speed-dial-actions.component.spec.ts
@@ -1,5 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
+import { PipFabSpeedDialComponent } from './fab-speed-dial.component';
 import { PipFabSpeedDialActionsComponent } from './fab-speed-dial-actions.component';
 
 describe('PipFabSpeedDialActionsComponent', () => {
@@ -8,9 +9,15 @@ describe('PipFabSpeedDialActionsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ PipFabSpeedDialActionsComponent ]
+      declarations: [PipFabSpeedDialActionsComponent],
+      providers: [
+        {
+          provide: PipFabSpeedDialComponent,
+          useValue: {}
+        }
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {