+++ src/app/fab-speed-dial-example/fab-speed-dial-example.component.spec.ts
@@ -1,4 +1,15 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import {
+  MatButtonModule,
+  MatButtonToggleModule,
+  MatCardModule,
+  MatIconModule,
+  MatRadioModule,
+  MatSlideToggleModule
+} from '@angular/material';
+import { PipFabSpeedDialModule } from 'pip-webui2-buttons';
 
 import { FabSpeedDialExampleComponent } from './fab-speed-dial-example.component';
 
@@ -8,9 +19,22 @@ describe('FabSpeedDialExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ FabSpeedDialExampleComponent ]
+      declarations: [FabSpeedDialExampleComponent],
+      imports: [
+        CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatButtonModule,
+        MatCardModule,
+        MatSlideToggleModule,
+        MatRadioModule,
+        MatIconModule,
+        MatButtonToggleModule,
+
+        PipFabSpeedDialModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {